bot/requirements.txt
anyio==4.2.0
APScheduler==3.10.4
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==20.7
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tzlocal==5.2
urllib3==2.1.0
